Add files via upload
Most of the lines have notes of what each one does but this code should be able to draw the board and place the white/black pieces accurately. The pieces can be dragged to all throughout the entire board. The only "rule" I've done is that we cannot take out the piece we've clicked on and each player gets a turn.